added deprecation for prev versions
